model/requirements.txt
Pillow
comet_ml
matplotlib
pydot
graphviz
git+https://github.com/tensorflow/examples.git
